Integrate CSS selector GUI
